+++ Test/Unit/Model/Campaign/Action/SendPushTest.php
@@ -7,6 +7,7 @@ use Ordo\Automation\Helper\Config;
 use Ordo\Automation\Model\Campaign\Action\SendPush;
 use Ordo\Automation\Model\Campaign\Action\SendRetrier;
 use Ordo\Automation\Model\Campaign\FrequencyCapGate;
+use Ordo\Automation\Model\Campaign\QuietHoursGate;
 use Ordo\Automation\Model\ConsentChannel;
 use Ordo\Automation\Model\ConsentManager;
 use Ordo\Automation\Model\Push\Exception\SubscriptionGoneException;
@@ -25,6 +26,7 @@ class SendPushTest extends TestCase
     private Config $config;
     private MessageLogWriter $messageLogWriter;
     private ConsentManager $consentManager;
+    private QuietHoursGate $quietHoursGate;
     private FrequencyCapGate $frequencyCapGate;
     private LoggerInterface $logger;
 
@@ -37,6 +39,8 @@ class SendPushTest extends TestCase
         $this->messageLogWriter = $this->createMock(MessageLogWriter::class);
         $this->consentManager = $this->createStub(ConsentManager::class);
         $this->consentManager->method('hasConsent')->willReturn(true);
+        $this->quietHoursGate = $this->createStub(QuietHoursGate::class);
+        $this->quietHoursGate->method('allows')->willReturn(true);
         $this->frequencyCapGate = $this->createStub(FrequencyCapGate::class);
         $this->frequencyCapGate->method('allows')->willReturn(true);
         $this->logger = $this->createMock(LoggerInterface::class);
@@ -50,6 +54,7 @@ class SendPushTest extends TestCase
             $this->config,
             $this->messageLogWriter,
             $this->consentManager,
+            $this->quietHoursGate,
             $this->frequencyCapGate,
             new SendRetrier(1),
             $this->logger
@@ -121,6 +126,20 @@ class SendPushTest extends TestCase
         $this->pushSubscriptionManager->expects(self::never())->method('getForCustomer');
         $this->pushSender->expects(self::never())->method('send');
         $this->messageLogWriter->expects(self::once())->method('recordOptedOut')->with('push', 42, '');
+
+        $context = ['customer_id' => 42];
+        $this->makeAction()->execute($context, ['title' => 'Hi']);
+    }
+
+    #[AllowMockObjectsWithoutExpectations]
+    public function testExecuteSkipsWhenQuietHoursGateDefers(): void
+    {
+        $this->quietHoursGate = $this->createMock(QuietHoursGate::class);
+        $this->quietHoursGate->expects(self::once())->method('allows')->willReturn(false);
+        $this->frequencyCapGate = $this->createMock(FrequencyCapGate::class);
+        $this->frequencyCapGate->expects(self::never())->method('allows');
+        $this->pushSubscriptionManager->expects(self::never())->method('getForCustomer');
+        $this->pushSender->expects(self::never())->method('send');
 
         $context = ['customer_id' => 42];
         $this->makeAction()->execute($context, ['title' => 'Hi']);